requirements.txt
chardet==5.0.0
numpy~=1.23.4
PyYAML~=6.0